// backend/src/exercises/interfaces/muscle-group.enum.ts
export enum MuscleGroupType {
  CHEST = 'CHEST',
  BACK = 'BACK',
  SHOULDERS = 'SHOULDERS',
  BICEPS = 'BICEPS',
  TRICEPS = 'TRICEPS',
  FOREARMS = 'FOREARMS',
  ABS = 'ABS',
  QUADRICEPS = 'QUADRICEPS',
  HAMSTRINGS = 'HAMSTRINGS',
  GLUTES = 'GLUTES',
  CALVES = 'CALVES',
  FULL_BODY = 'FULL_BODY',
  CARDIO = 'CARDIO',
  OTHER = 'OTHER',
}